# app/Http/Requests/API/POS/ConnectDeviceRequest.php
<?php

namespace App\Http\Requests\API\POS;

use App\Http\Requests\API\V1\POS\ConnectDeviceRequest as V1ConnectDeviceRequest;

/**
 * @deprecated Legacy unversioned request. Please use App\Http\Requests\API\V1\POS\ConnectDeviceRequest.
 */
class ConnectDeviceRequest extends V1ConnectDeviceRequest {}