# app/Http/Requests/API/POS/OpenShiftRequest.php
<?php

namespace App\Http\Requests\API\POS;

use App\Http\Requests\API\V1\POS\OpenShiftRequest as V1OpenShiftRequest;

/**
 * @deprecated Legacy unversioned request. Please use App\Http\Requests\API\V1\POS\OpenShiftRequest.
 */
class OpenShiftRequest extends V1OpenShiftRequest {}