# app/Http/Requests/API/POS/StoreCashLogRequest.php
<?php

namespace App\Http\Requests\API\POS;

use App\Http\Requests\API\V1\POS\StoreCashLogRequest as V1StoreCashLogRequest;

/**
 * @deprecated Legacy unversioned request. Please use App\Http\Requests\API\V1\POS\StoreCashLogRequest.
 */
class StoreCashLogRequest extends V1StoreCashLogRequest {}