# app/Http/Requests/API/POS/StorePosTransactionRequest.php
<?php

namespace App\Http\Requests\API\POS;

use App\Http\Requests\API\V1\POS\StorePosTransactionRequest as V1StorePosTransactionRequest;

/**
 * @deprecated Legacy unversioned request. Please use App\Http\Requests\API\V1\POS\StorePosTransactionRequest.
 */
class StorePosTransactionRequest extends V1StorePosTransactionRequest {}