# app/Http/Requests/API/V1/POS/OpenShiftRequest.php
<?php

namespace App\Http\Requests\API\V1\POS;

use Illuminate\Foundation\Http\FormRequest;

class OpenShiftRequest extends FormRequest
{
    public function authorize(): bool
    {
        return true;
    }

    public function rules(): array
    {
        return [
            'user_id' => ['required', 'string', 'uuid'],
            'opening_cash' => ['required', 'numeric', 'min:0'],
        ];
    }
}