# docs/openapi.yaml
openapi: 3.0.3
info:
    title: Sollu POS Client API
    version: 1.0.0
    description: API Spesifikasi untuk Aplikasi POS Client (Flutter) & Pairing Perangkat
servers:
    - url: http://api.sollu.test/v1/pos
      description: Local V1 Server (Canonical)
    - url: http://api.sollu.test/pos
      description: Local Fallback Server (Deprecated)
    - url: http://app.sollu.test/api/pos
      description: Legacy App Bridge Server (Deprecated)
components:
    securitySchemes:
        SanctumAuth:
            type: http
            scheme: bearer
            bearerFormat: Sanctum Token
    parameters:
        DeviceUuidHeader:
            name: X-DEVICE-UUID
            in: header
            required: true
            schema:
                type: string
                format: uuid
            example: 'f81d4fae-7dec-11d0-a765-00a0c91e6bf6'
        HardwareSignatureHeader:
            name: X-HARDWARE-SIGNATURE
            in: header
            required: true
            schema:
                type: string
            example: 'a3f8c...'
    schemas:
        ConnectDeviceRequest:
            type: object
            required:
                - otp
                - device_uuid
                - hardware_fingerprint
            properties:
                otp:
                    type: string
                    example: '12345678'
                    description: Kode OTP 8-digit yang dihasilkan dari Dashboard
                device_uuid:
                    type: string
                    format: uuid
                    example: 'f81d4fae-7dec-11d0-a765-00a0c91e6bf6'
                    description: UUID spesifik perangkat dari aplikasi POS
                hardware_fingerprint:
                    type: string
                    example: 'a3f8c...'
                    description: Signature/fingerprint hardware dari aplikasi POS
        ConnectDeviceResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: Perangkat berhasil dihubungkan.
                data:
                    type: object
                    properties:
                        token:
                            type: string
                            example: '1|abcdef123456...'
                        device:
                            type: object
                            properties:
                                id:
                                    type: string
                                    format: uuid
                                name:
                                    type: string
                                    example: 'Kasir Utama'
                                type:
                                    type: string
                                    example: 'pos'
                        outlet:
                            type: object
                            properties:
                                id:
                                    type: string
                                    format: uuid
                                name:
                                    type: string
                                    example: 'Outlet Sudirman'
                                address:
                                    type: string
                                    example: 'Jl. Jend. Sudirman No. 1'
                                phone_number:
                                    type: string
                                    example: '081234567890'
                        business:
                            type: object
                            properties:
                                id:
                                    type: string
                                    format: uuid
                                name:
                                    type: string
                                    example: 'Sollu Coffee & Resto'

        StorePosTransactionRequest:
            type: object
            required:
                - offline_id
                - subtotal
                - discount_amount
                - tax_amount
                - service_charge_amount
                - total
                - payment_status
                - status
                - items
            properties:
                offline_id:
                    type: string
                    format: uuid
                    example: '9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d'
                receipt_number:
                    type: string
                    example: 'INV/20260810/0001'
                customer_id:
                    type: string
                    format: uuid
                    nullable: true
                subtotal:
                    type: number
                    example: 50000
                discount_amount:
                    type: number
                    example: 5000
                tax_amount:
                    type: number
                    example: 4950
                service_charge_amount:
                    type: number
                    example: 0
                total:
                    type: number
                    example: 49950
                payment_status:
                    type: string
                    enum: [unpaid, paid, partial]
                    example: 'paid'
                status:
                    type: string
                    enum: [hold, completed, void]
                    example: 'completed'
                notes:
                    type: string
                    example: 'Catatan khusus pesanan'
                items:
                    type: array
                    items:
                        type: object
                        required:
                            - product_id
                            - product_name
                            - price
                            - qty
                            - discount_amount
                            - subtotal
                        properties:
                            product_id:
                                type: string
                                format: uuid
                            variant_group_option_id:
                                type: string
                                format: uuid
                                nullable: true
                            product_name:
                                type: string
                                example: 'Kopi Susu Gula Aren'
                            price:
                                type: number
                                example: 25000
                            qty:
                                type: integer
                                example: 2
                            discount_amount:
                                type: number
                                example: 0
                            subtotal:
                                type: number
                                example: 50000
                            modifiers:
                                type: array
                                items:
                                    type: object
                                    required:
                                        - modifier_option_id
                                        - modifier_name
                                        - price
                                        - qty
                                    properties:
                                        modifier_option_id:
                                            type: string
                                            format: uuid
                                        modifier_name:
                                            type: string
                                            example: 'Extra Shot Espresso'
                                        price:
                                            type: number
                                            example: 5000
                                        qty:
                                            type: integer
                                            example: 1
                payments:
                    type: array
                    items:
                        type: object
                        required:
                            - payment_method_id
                            - amount
                            - change_amount
                        properties:
                            payment_method_id:
                                type: string
                                format: uuid
                            amount:
                                type: number
                                example: 50000
                            change_amount:
                                type: number
                                example: 50
                            payment_reference:
                                type: string
                                nullable: true

        OpenShiftRequest:
            type: object
            required:
                - user_id
                - opening_cash
            properties:
                user_id:
                    type: string
                    format: uuid
                opening_cash:
                    type: number
                    example: 500000

        CloseShiftRequest:
            type: object
            required:
                - closing_cash
                - actual_cash
            properties:
                closing_cash:
                    type: number
                    example: 1250000
                actual_cash:
                    type: number
                    example: 1250000

        StoreCashLogRequest:
            type: object
            required:
                - type
                - amount
            properties:
                type:
                    type: string
                    enum: [cash_in, cash_out]
                    example: 'cash_out'
                amount:
                    type: number
                    example: 50000
                description:
                    type: string
                    description: 'Beli es batu kristal'

        HealthCheckResponse:
            type: object
            properties:
                status:
                    type: string
                    example: 'healthy'
                timestamp:
                    type: string
                    example: '2026-08-29T02:50:48+07:00'
                environment:
                    type: string
                    example: 'production'
                services:
                    type: object
                    properties:
                        database:
                            type: object
                            properties:
                                status:
                                    type: string
                                    example: 'healthy'
                                latency_ms:
                                    type: number
                                    example: 1.25
                        cache:
                            type: object
                            properties:
                                status:
                                    type: string
                                    example: 'healthy'
                                driver:
                                    type: string
                                    example: 'redis'
                        storage:
                            type: object
                            properties:
                                status:
                                    type: string
                                    example: 'healthy'

        MidtransNotificationRequest:
            type: object
            description: Payload notifikasi dari Midtrans (contoh field utama)
            properties:
                transaction_time:
                    type: string
                    example: '2026-08-10 15:00:00'
                transaction_status:
                    type: string
                    example: 'settlement'
                transaction_id:
                    type: string
                    example: '5409...-...'
                status_message:
                    type: string
                    example: 'midtrans payment notification'
                status_code:
                    type: string
                    example: '200'
                signature_key:
                    type: string
                    example: '9c...e0'
                payment_type:
                    type: string
                    example: 'bank_transfer'
                order_id:
                    type: string
                    example: 'INV-12345'
                merchant_id:
                    type: string
                    example: 'G...1'
                gross_amount:
                    type: string
                    example: '150000.00'
                fraud_status:
                    type: string
                    example: 'accept'


paths:
    /health:
        get:
            summary: Health Check Endpoint
            description: Endpoint pemantauan status kesehatan aplikasi, koneksi database, cache/redis, dan storage.
            responses:
                '200':
                    description: Sistem berjalan normal (healthy)
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HealthCheckResponse'
                '503':
                    description: Satu atau lebih dependensi kritis sistem mengalami gangguan (unhealthy)
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HealthCheckResponse'

    /device/connect:
        post:
            summary: Connect Device & Pairing POS
            description: Memvalidasi 8-digit OTP, mendaftarkan device_uuid & hardware_fingerprint dari POS Client, serta mengembalikan Sanctum Bearer Token.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ConnectDeviceRequest'
            responses:
                '200':
                    description: Berhasil verifikasi dan mendapatkan token
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ConnectDeviceResponse'
                '400':
                    description: OTP Tidak Valid / Kadaluarsa

    /device/status:
        get:
            summary: Cek Status Koneksi Device
            description: Memeriksa apakah token Sanctum dan fingerprint hardware masih valid/aktif.
            security:
                - SanctumAuth: []
            parameters:
                - $ref: '#/components/parameters/DeviceUuidHeader'
                - $ref: '#/components/parameters/HardwareSignatureHeader'
            responses:
                '200':
                    description: Device valid dan terhubung
                '401':
                    description: Token atau signature tidak valid

    /sync/master:
        get:
            summary: Pull Master Data
            description: Menarik seluruh master data toko (Produk, Kategori, Metode Pembayaran, Setting Outlet) untuk disimpan ke SQLite POS.
            security:
                - SanctumAuth: []
            parameters:
                - $ref: '#/components/parameters/DeviceUuidHeader'
                - $ref: '#/components/parameters/HardwareSignatureHeader'
            responses:
                '200':
                    description: Payload Master Data berhasil diambil

    /employees:
        get:
            summary: Get Employees List
            description: Menarik daftar karyawan/kasir yang terdaftar pada outlet tempat device dipairing.
            security:
                - SanctumAuth: []
            parameters:
                - $ref: '#/components/parameters/DeviceUuidHeader'
                - $ref: '#/components/parameters/HardwareSignatureHeader'
            responses:
                '200':
                    description: Data karyawan berhasil diambil
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    message:
                                        type: string
                                        example: 'Data karyawan berhasil diambil.'
                                    data:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                id:
                                                    type: string
                                                    format: uuid
                                                name:
                                                    type: string
                                                    example: 'Budi Kasir'
                                                email:
                                                    type: string
                                                    example: 'budi@sollu.id'
                                                pin:
                                                    type: string
                                                    example: '$2y$12$eImiTXuWVxfM37uY4JANjO...'
                                                photo:
                                                    type: string
                                                    nullable: true
                                                role:
                                                    type: string
                                                    example: 'Kasir'

    /transactions:
        post:
            summary: Sync Offline Transaction (Push)
            description: Mengirim data transaksi dari SQLite lokal POS ke server cloud. Menggunakan idempotency check pada offline_id.
            security:
                - SanctumAuth: []
            parameters:
                - $ref: '#/components/parameters/DeviceUuidHeader'
                - $ref: '#/components/parameters/HardwareSignatureHeader'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StorePosTransactionRequest'
            responses:
                '200':
                    description: Transaksi berhasil disinkronisasi

    /shifts/open:
        post:
            summary: Buka Shift Kasir
            security:
                - SanctumAuth: []
            parameters:
                - $ref: '#/components/parameters/DeviceUuidHeader'
                - $ref: '#/components/parameters/HardwareSignatureHeader'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/OpenShiftRequest'
            responses:
                '200':
                    description: Shift berhasil dibuka

    /shifts/close:
        post:
            summary: Tutup Shift Kasir
            security:
                - SanctumAuth: []
            parameters:
                - $ref: '#/components/parameters/DeviceUuidHeader'
                - $ref: '#/components/parameters/HardwareSignatureHeader'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CloseShiftRequest'
            responses:
                '200':
                    description: Shift berhasil ditutup

    /shifts/cash-log:
        post:
            summary: Catat Cash In / Cash Out
            security:
                - SanctumAuth: []
            parameters:
                - $ref: '#/components/parameters/DeviceUuidHeader'
                - $ref: '#/components/parameters/HardwareSignatureHeader'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StoreCashLogRequest'
            responses:
                '200':
                    description: Log kas berhasil dibuat

    /midtrans/notification:
        post:
            summary: Midtrans Webhook Notification
            description: Endpoint untuk menerima HTTP callback (webhook) dari server Midtrans ketika status pembayaran berubah (misal success, failed, expire).
            servers:
                - url: http://app.sollu.test/api
                  description: Base API Server
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MidtransNotificationRequest'
            responses:
                '200':
                    description: Webhook berhasil diproses (termasuk kasus payment already processed)
                '403':
                    description: Signature key tidak valid (Kemungkinan percobaan spoofing)
                '404':
                    description: Payment record (Order ID) tidak ditemukan di database